Move builders outside AnnotationSuppressorSpec
